moved out logging
